fix(plugin-ovn): support 4-byte BGP ASNs in the /32 announce path
The ASN chain (BgpFrrAsn ConfigKey, OvnBgpAnnounceCommand/Answer, and
the agent-side auto-detect in LibvirtOvnBgpAnnounceCommandWrapper) used
Integer end to end. eBGP fabrics using RFC 6996 4-byte private ASNs
(4200000000-4294967294, one per host) overflow Integer.parseInt, so
auto-detect silently returned null and every public-IP /32 announce
failed with "ASN auto-detect failed" — static NAT and source NAT IPs
never became reachable. The whole chain is now Long.

+++ api/src/main/java/com/cloud/agent/api/OvnBgpAnnounceAnswer.java
@@ -24,7 +24,7 @@ package com.cloud.agent.api;
  */
 public class OvnBgpAnnounceAnswer extends Answer {
 
-    private Integer asn;
+    private Long asn;
 
     /** No-arg constructor for serialization frameworks. */
     public OvnBgpAnnounceAnswer() {
@@ -35,16 +35,16 @@ public class OvnBgpAnnounceAnswer extends Answer {
         super(command, success, details);
     }
 
-    public OvnBgpAnnounceAnswer(final Command command, final boolean success, final String details, final Integer asn) {
+    public OvnBgpAnnounceAnswer(final Command command, final boolean success, final String details, final Long asn) {
         super(command, success, details);
         this.asn = asn;
     }
 
-    public Integer getAsn() {
+    public Long getAsn() {
         return asn;
     }
 
-    public void setAsn(final Integer asn) {
+    public void setAsn(final Long asn) {
         this.asn = asn;
     }
 }
